CRUD working fine(category-management v0.4)

+++ src/main/java/com/itacademy/cms/service/impl/CategoryServiceImpl.java
@@ -1,9 +1,14 @@
 package com.itacademy.cms.service.impl;
 
-import com.itacademy.cms.jpaRepository.CategoryRepository;
+import com.itacademy.cms.exeption.CategoryNotFoundException;
+import com.itacademy.cms.exeption.ParameterMissingException;
+import com.itacademy.cms.mapper.CategoryMapper;
 import com.itacademy.cms.model.Category;
+import com.itacademy.cms.model.dto.CategoryDto;
+import com.itacademy.cms.repository.CategoryRepository;
 import com.itacademy.cms.service.CategoryService;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,20 +18,50 @@ public class CategoryServiceImpl implements CategoryService {
 
 
   private final CategoryRepository categoryRepository;
+  private final CategoryMapper categoryMapper;
 
 
   @Override
-  public List<Category> getAllCategories() {
-    return categoryRepository.findAll();
+  public List<Category> findAll() throws CategoryNotFoundException {
+    List<Category> categoriesList = categoryRepository.findAll();
+    if (categoriesList.isEmpty()) {
+      throw new CategoryNotFoundException("No categories found!");
+    }
+    return categoriesList;
   }
 
   @Override
-  public void saveCategory(Category category) {
-    categoryRepository.save(category);
+  public void updateCategory(CategoryDto categoryDto, Long id) throws CategoryNotFoundException {
+    Optional<Category> categoryOptional = categoryRepository.findById(id);
+    categoryOptional.ifPresent(category -> {
+      category.setCourses(categoryDto.getCourses());
+      category.setCategoryName(categoryDto.getCategoryName());
+      categoryRepository.save(category);
+    });
   }
 
   @Override
-  public void deleteCategory(int id) {
-    categoryRepository.deleteById(id);
+  public Category findById(Long id) throws CategoryNotFoundException {
+    Optional<Category> category = categoryRepository.findById(id);
+    if (category.isPresent()) {
+      return categoryRepository.getById(id);
+    }
+    throw new CategoryNotFoundException("Category with id " + id + " not found!");
   }
+
+  @Override
+  public void saveCategory(CategoryDto categoryDto) {
+    categoryRepository.save(categoryMapper.categoryDtoToCategory(categoryDto));
+  }
+
+  @Override
+  public void deleteCategoryById(Long id) throws CategoryNotFoundException {
+    if (id == null) {
+      throw new ParameterMissingException("Category id is missing");
+    } else if (categoryRepository.existsById(id)) {
+      categoryRepository.deleteById(id);
+    }
+    throw new CategoryNotFoundException("Category with id " + id + " not found!");
+  }
+
 }
